Them chuc nang Add User & Delete User

+++ frontend/src/components/TaskList.vue
@@ -5,34 +5,36 @@
                 <th>STT</th>
                 <th>Decriptions</th>
                 <th>Status</th>
-                <th>Delete?</th>               
-            
+                <th>Delete?</th>
+
             </tr>
         </thead>
 
         <tbody>
             <tr v-for="(task, index) in taskLists" :key="task.id">
-                <td>{{index + 1}}</td>
-                <td>{{task.description}}</td>
-                <td><input type="checkbox" v-model="task.isComplete" name="" id="" @click="completeTask(task._id)">Done</td>
-                <td><button @click="deleteTask(task._id)"><i class="fa-solid fa-arrow-left"></i>Delete this task<i class="fa-solid fa-trash-can"></i></button></td>
-            </tr>        
+                <td>{{ index + 1 }}</td>
+                <td>{{ task.description }}</td>
+                <td><input type="checkbox" v-model="task.isComplete" name="" id="" @click="completeTask(task._id)">Done
+                </td>
+                <td><button @click="deleteTask(task._id)"><i class="fa-solid fa-arrow-left"></i>Delete this task<i
+                            class="fa-solid fa-trash-can"></i></button></td>
+            </tr>
         </tbody>
-        
+
         <h5 style="color:green">
-            Progress:{{(totalDone / (taskLists.length > 0 ? taskLists.length : 1) * 100).toFixed(2)}}%
+            Progress:{{ (totalDone / (taskLists.length > 0 ? taskLists.length : 1) * 100).toFixed(2) }}%
         </h5>
 
-        <button class="btn btn-primary"  @click="goToAddTask">Assign new task<i class="fa-solid fa-plus"></i></button>
-        
+        <button class="btn btn-primary" @click="goToAddTask">Assign new task<i class="fa-solid fa-plus"></i></button>
+
     </table>
 </template>
 <script>
 import TaskService from '../services/task.service';
-export default{
+export default {
     data() {
         return {
-            totalDone:0,
+            totalDone: 0,
             taskLists: []
         };
     },
@@ -40,12 +42,12 @@ export default{
         let userId = this.$route.params.id;
         this.getTask(userId);
     },
-    methods:{
+    methods: {
         goToAddTask() {
             this.$router.push({ name: "task.add" });
         },
-        async completeTask(id){
-            console.log('Update isComplete field of TaskId: ',  id)
+        async completeTask(id) {
+            console.log('Update isComplete field of TaskId: ', id)
             try {
                 await TaskService.update(id);
                 this.updateTotalDone();
@@ -53,17 +55,18 @@ export default{
                 console.log(error);
             }
         },
-        updateTotalDone(){
+        updateTotalDone() {
             this.totalDone = 0;
             this.taskLists.forEach((task) => {
-                if(task.isComplete) {
+                if (task.isComplete) {
                     this.totalDone++;
                 }
             })
         },
         async deleteTask(id) {
-            try {
-                // console.log('Delete TaskId: ',id)
+            if (confirm("Bạn muốn xóa Task này?")) {
+                try {
+                    // console.log('Delete TaskId: ',id)
                     await TaskService.delete(id);
                     // refreshList
                     let userId = this.$route.params.id;
@@ -72,8 +75,10 @@ export default{
                 } catch (error) {
                     console.log(error);
                 }
+            }
         },
-        async getTask(id){
+
+        async getTask(id) {
             this.taskLists = await TaskService.getByUserId(id);
             this.updateTotalDone();
         },
